refactor(tinyusb): Renamed tinyusb default config header file
