+++ device/esp_tinyusb/test_apps/cdc/main/test_cdc.c
@@ -17,7 +17,7 @@
 
 #include "unity.h"
 #include "tinyusb.h"
-#include "tinyusb_default_configs.h"
+#include "tinyusb_default_config.h"
 #include "tusb_cdc_acm.h"
 #include "vfs_tinyusb.h"
