+++ device/esp_tinyusb/test_apps/msc_storage/main/test_msc_storage.c
@@ -22,7 +22,7 @@
 #include "unity.h"
 #include "device_common.h"
 #include "tinyusb.h"
-#include "tinyusb_default_configs.h"
+#include "tinyusb_default_config.h"
 #include "tinyusb_msc.h"
 #include "storage_common.h"
